# Conflicts:
#	Assets/Imported Assets/Beautify/URP.meta
#	Assets/Imported Assets/Bitgem/DoTween/Plugins/Demigiant.meta
#	Assets/Imported Assets/Bitgem/DoTween/Plugins/Demigiant/DOTween.meta
#	Assets/Imported Assets/Bitgem/DoTween/Plugins/Demigiant/DOTween/Editor.meta
#	Assets/Imported Assets/Bitgem/DoTween/Plugins/Demigiant/DOTween/Editor/Imgs.meta
#	Assets/Imported Assets/DunGen/DoTween/Core/Editor.meta
#	Assets/Imported Assets/DunGen/DoTween/Core/Editor/Imgs.meta
#	Assets/Imported Assets/DunGen/DoTween/Core/Editor/Imgs/project.meta
#	Assets/Plugins/Easy performant outline/Legacy version.meta
#	Assets/Resources/Prefab/UI/SimpleInput/Scripts.meta
#	Assets/Scripts/Unit/LivingEntity/Monster/MonsterAction/Player.meta

+++ Assets/Scripts/Unit/LivingEntity.cs
@@ -42,7 +42,7 @@ public class LivingEntity : Unit
     public virtual void Damaged(float damage)
     {
         _hp -= damage;
-        ObjectPoolManager.Instance.GetObject(DamageText, transform.position, Quaternion.identity).GetComponent<DamageText>().PlayDamage(damage, true);
+        ObjectPoolManager.Instance.GetObject(DamageText, transform.position, Quaternion.identity).GetComponent<DamageText>().PlayDamage((int)damage, true);
     }
 
     public virtual void UseStemina(float skillmp)
